chore: remove hard set tap colors
The original PR https://github.com/npm/cli/pull/2225 indicated that this
would prevent snapshot failures in CI with no color, but I hard set this
to "0" and ran tests and they passed.  This will make debugging tests in
CI environments like the Node.js CITGM much easier.

The reason for the failures is because even if we hard set environment
variables in tests to force color one way or another, chalk has already
made up its mind so changing those in the same process won't affect
anything.  The fix is to hard set two values in the actual `npm test`
script to make `chalk` and `supports-color` do what we need.

The `sudo` test commands were also removed because we do not want to
encourage running npm w/ sudo.